+++ src/main.ts
@@ -4,15 +4,15 @@ import { toggleStopAudioWhenReleased } from './audio/stopAudioWhenReleased.ts';
 import { toggleKeyboardMode } from './audio/switchKeyboard.ts';
 import { toggleLights } from './visual/lights.ts';
 import { octaveUp, octaveDown, transposeUpOne, transposeDownOne } from './audio/transposeOctave.ts';
-import { showControlsButton, toggleModal } from './components/modal.ts';
+import { hideAllModals, toggleModal } from './components/modal.ts';
 import './audio/recording.ts';
 import './core/cumTime.ts';
 import { keyEventToBaseKey, registerKeyDown, registerKeyUp } from './core/keypress.ts';
 import { states } from './core/states.ts';
-// import './sheets/transcribe.ts';
 import './sheets/sheetPlayer.ts'
 // import changelogMD from '/CHANGELOG.md?raw';
 // import showdown from 'showdown';
+
 
 
 
@@ -27,44 +27,7 @@ document.addEventListener("DOMContentLoaded", () => {
 
     // ===========================================
 
-    function isFirstVisit(): boolean {
-        if (!localStorage.getItem('visited')) {
-          localStorage.setItem('visited', 'true');
-          return true;
-        }
-        // else
-        return false;
-      }
-      
-      if (isFirstVisit()) {
-        // first time
-        document.getElementById("skysynth-greeting")!.textContent = "hello"
-        showControlsButton.dispatchEvent(new Event("pointerdown"));
-      } else {  
-        // repeated visit
-        document.getElementById("skysynth-greeting")!.textContent = "welcome back"
-        if (isFirstVisitAfterUpdate()) {
-          // // show modal that shows most recent update
-          // const changelogModalContent = document.getElementById("changelog-modal-content")!;
-          // // convert changelog md to html
-          // showdown.setFlavor('github');
-          // const showdownConverter = new showdown.Converter();
-          // const showdownText = changelogMD;
-          // const showDdownHTML = showdownConverter.makeHtml(showdownText);
-          // changelogModalContent.innerHTML = showDdownHTML;
-          // toggleModal(true, "changelog");
-        }
-      }
 
-      function isFirstVisitAfterUpdate(): boolean {
-        if (localStorage.getItem('lastVisitedVersion') != states.skysynthVersion) {
-          states.skysynthVersionOnLastVisit = localStorage.getItem('lastVisitedVersion');
-          localStorage.setItem('lastVisitedVersion', states.skysynthVersion);
-          return true;
-        }
-        // else
-        return false;
-      }
 
 
     // ===========================================
@@ -89,8 +52,9 @@ document.addEventListener("DOMContentLoaded", () => {
     document.getElementById("modal-alt-r")!.addEventListener("pointerup", () => document.getElementById("r-alt-indicator")!.style.backgroundColor = "");
     document.getElementById("modal-capslock")!.addEventListener("pointerdown", () => toggleStopAudioWhenReleased());
     document.getElementById("modal-\\")!.addEventListener("pointerdown", toggleLights);
-    document.getElementById("modal-tab")!.addEventListener("pointerdown", () => toggleModal(false));
-    document.getElementById("modal-esc")!.addEventListener("pointerdown", () => toggleModal(false));
+    document.getElementById("modal-tab")!.addEventListener("pointerdown", () => {hideAllModals(); toggleModal(true, "transcribe")});
+    document.getElementById("modal-delete")!.addEventListener("pointerdown", () => {hideAllModals(); toggleModal(true, "debug")});
+    document.getElementById("modal-esc")!.addEventListener("pointerdown", () => hideAllModals);
 
     // ===========================================
     // HANDLING KEYPRESSES
@@ -102,5 +66,3 @@ document.addEventListener("DOMContentLoaded", () => {
     updateStatusMsg("Initialised!");
 
 })
-
-// export { volumeNode };